Adjust repository package layout and ensure confg classes are generic enough
