+++ dapr-spring-core/src/main/java/io/diagrid/spring/core/keyvalue/PostgreSQLDaprKeyValueAdapter.java
@@ -16,7 +16,7 @@ package io.diagrid.spring.core.keyvalue;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.dapr.client.DaprClient;
 import io.dapr.utils.TypeRef;
-import io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator;
+import io.diagrid.spring.core.keyvalue.repository.query.DaprPredicateQueryCreator;
 import org.springframework.data.keyvalue.core.query.KeyValueQuery;
 import org.springframework.expression.spel.SpelNode;
 import org.springframework.expression.spel.standard.SpelExpression;
@@ -138,11 +138,11 @@ public class PostgreSQLDaprKeyValueAdapter extends AbstractDaprKeyValueAdapter {
 
   private String createSql(String sqlPattern, String keyspace, KeyValueQuery<?> query) {
     String keyspaceFilter = getKeyspaceFilter(keyspace);
-    if (query.getCriteria() instanceof PredicateQueryCreator.ValueComparingPredicate) {
-      String path = ((PredicateQueryCreator.ValueComparingPredicate) query.getCriteria()).getPath().toString();
-      String pathWithOutType = String.format("'%s'", path.substring(path.indexOf(".") + 1, path.length()));
+    if (query.getCriteria() instanceof DaprPredicateQueryCreator.ValueComparingPredicate) {
+      String path = ((DaprPredicateQueryCreator.ValueComparingPredicate) query.getCriteria()).getPath().toString();
+      String pathWithOutType = String.format("'%s'", path.substring(path.indexOf(".") + 1));
       String value = String.format("'%s'",
-          ((PredicateQueryCreator.ValueComparingPredicate) query.getCriteria()).getValue().toString());
+          ((DaprPredicateQueryCreator.ValueComparingPredicate) query.getCriteria()).getValue().toString());
       return String.format(sqlPattern, keyspaceFilter, pathWithOutType, value);
     } else if (query.getCriteria() instanceof String) {
       SpelExpression expression = PARSER.parseRaw(query.getCriteria().toString());
