+++ dapr-spring-core/src/main/java/io/diagrid/spring/core/keyvalue/repository/config/DaprRepositoriesRegistrar.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2024 The Dapr Authors
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package io.diagrid.spring.core.keyvalue.repository.config;
+
+import org.springframework.data.repository.config.RepositoryBeanDefinitionRegistrarSupport;
+import org.springframework.data.repository.config.RepositoryConfigurationExtension;
+
+import java.lang.annotation.Annotation;
+
+/**
+ * Dapr specific {@link RepositoryBeanDefinitionRegistrarSupport} implementation.
+ */
+public class DaprRepositoriesRegistrar extends RepositoryBeanDefinitionRegistrarSupport {
+
+  @Override
+  protected Class<? extends Annotation> getAnnotation() {
+    return EnableDaprRepositories.class;
+  }
+
+  @Override
+  protected RepositoryConfigurationExtension getExtension() {
+    return new DaprRepositoryConfigurationExtension();
+  }
+}
