+++ dapr-spring-core/src/main/java/io/diagrid/spring/core/keyvalue/repository/query/DaprPredicateQueryCreator.java
@@ -11,7 +11,7 @@
 limitations under the License.
 */
 
-package io.diagrid.spring.core.keyvalue.repository;
+package io.diagrid.spring.core.keyvalue.repository.query;
 
 import org.springframework.dao.InvalidDataAccessApiUsageException;
 import org.springframework.data.domain.Sort;
@@ -39,11 +39,11 @@ import java.util.regex.Pattern;
  * This class is copied from https://github.com/spring-projects/spring-data-keyvalue/blob/ff441439124585042dd0cbff952f977a343444d2/src/main/java/org/springframework/data/keyvalue/repository/query/PredicateQueryCreator.java#L46
  * because it has private accessors to internal classes, making it impossible to extend or use the original
  * This requires to be created from scratch to not use predicates, but this is only worth it if we can prove these
- * abstractions are worth the time
+ * abstractions are worth the time.
  */
-public class PredicateQueryCreator extends AbstractQueryCreator<KeyValueQuery<Predicate<?>>, Predicate<?>> {
+public class DaprPredicateQueryCreator extends AbstractQueryCreator<KeyValueQuery<Predicate<?>>, Predicate<?>> {
 
-  public PredicateQueryCreator(PartTree tree, ParameterAccessor parameters) {
+  public DaprPredicateQueryCreator(PartTree tree, ParameterAccessor parameters) {
     super(tree, parameters);
   }
 
@@ -51,53 +51,53 @@ public class PredicateQueryCreator extends AbstractQueryCreator<KeyValueQuery<Pr
   protected Predicate<?> create(Part part, Iterator<Object> iterator) {
     switch (part.getType()) {
       case TRUE:
-        return io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
+        return DaprPredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
             .isTrue();
       case FALSE:
-        return io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
+        return DaprPredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
             .isFalse();
       case SIMPLE_PROPERTY:
-        return io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
+        return DaprPredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
             .isEqualTo(iterator.next());
       case IS_NULL:
-        return io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
+        return DaprPredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
             .isNull();
       case IS_NOT_NULL:
-        return io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
+        return DaprPredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
             .isNotNull();
       case LIKE:
-        return io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
+        return DaprPredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
             .contains(iterator.next());
       case STARTING_WITH:
-        return io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
+        return DaprPredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
             .startsWith(iterator.next());
       case AFTER:
       case GREATER_THAN:
-        return io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
+        return DaprPredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
             .isGreaterThan(iterator.next());
       case GREATER_THAN_EQUAL:
-        return io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
+        return DaprPredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
             .isGreaterThanEqual(iterator.next());
       case BEFORE:
       case LESS_THAN:
-        return io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
+        return DaprPredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
             .isLessThan(iterator.next());
       case LESS_THAN_EQUAL:
-        return io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
+        return DaprPredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
             .isLessThanEqual(iterator.next());
       case ENDING_WITH:
-        return io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
+        return DaprPredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
             .endsWith(iterator.next());
       case BETWEEN:
-        return io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
+        return DaprPredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
             .isGreaterThan(iterator.next())
-            .and(io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
+            .and(DaprPredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
                 .isLessThan(iterator.next()));
       case REGEX:
-        return io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
+        return DaprPredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
             .matches(iterator.next());
       case IN:
-        return io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
+        return DaprPredicateQueryCreator.PredicateBuilder.propertyValueOf(part)
             .in(iterator.next());
       default:
         throw new InvalidDataAccessApiUsageException(String.format("Found invalid part '%s' in query", part.getType()));
@@ -131,23 +131,23 @@ public class PredicateQueryCreator extends AbstractQueryCreator<KeyValueQuery<Pr
       this.part = part;
     }
 
-    static io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.PredicateBuilder propertyValueOf(
+    static DaprPredicateQueryCreator.PredicateBuilder propertyValueOf(
         Part part) {
-      return new io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.PredicateBuilder(part);
+      return new DaprPredicateQueryCreator.PredicateBuilder(part);
     }
 
     public Predicate<Object> isTrue() {
-      return new io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.ValueComparingPredicate(
+      return new DaprPredicateQueryCreator.ValueComparingPredicate(
           part.getProperty(), true);
     }
 
     public Predicate<Object> isFalse() {
-      return new io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.ValueComparingPredicate(
+      return new DaprPredicateQueryCreator.ValueComparingPredicate(
           part.getProperty(), false);
     }
 
     public Predicate<Object> isEqualTo(Object value) {
-      return new io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.ValueComparingPredicate(
+      return new DaprPredicateQueryCreator.ValueComparingPredicate(
           part.getProperty(), value, o -> {
 
         if (!ObjectUtils.nullSafeEquals(IgnoreCaseType.NEVER, part.shouldIgnoreCase())) {
@@ -161,7 +161,7 @@ public class PredicateQueryCreator extends AbstractQueryCreator<KeyValueQuery<Pr
     }
 
     public Predicate<Object> isNull() {
-      return new io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.ValueComparingPredicate(
+      return new DaprPredicateQueryCreator.ValueComparingPredicate(
           part.getProperty(), null, Objects::isNull);
     }
 
@@ -170,32 +170,32 @@ public class PredicateQueryCreator extends AbstractQueryCreator<KeyValueQuery<Pr
     }
 
     public Predicate<Object> isLessThan(Object value) {
-      return new io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.ValueComparingPredicate(
+      return new DaprPredicateQueryCreator.ValueComparingPredicate(
           part.getProperty(), value,
-          o -> NullSafeComparator.NULLS_HIGH.compare(o, value) == -1 ? true : false);
+          o -> NullSafeComparator.NULLS_HIGH.compare(o, value) == -1);
     }
 
     public Predicate<Object> isLessThanEqual(Object value) {
-      return new io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.ValueComparingPredicate(
+      return new DaprPredicateQueryCreator.ValueComparingPredicate(
           part.getProperty(), value,
-          o -> NullSafeComparator.NULLS_HIGH.compare(o, value) <= 0 ? true : false);
+          o -> NullSafeComparator.NULLS_HIGH.compare(o, value) <= 0);
     }
 
     public Predicate<Object> isGreaterThan(Object value) {
-      return new io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.ValueComparingPredicate(
+      return new DaprPredicateQueryCreator.ValueComparingPredicate(
           part.getProperty(), value,
-          o -> NullSafeComparator.NULLS_HIGH.compare(o, value) == 1 ? true : false);
+          o -> NullSafeComparator.NULLS_HIGH.compare(o, value) == 1);
     }
 
     public Predicate<Object> isGreaterThanEqual(Object value) {
-      return new io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.ValueComparingPredicate(
+      return new DaprPredicateQueryCreator.ValueComparingPredicate(
           part.getProperty(), value,
-          o -> NullSafeComparator.NULLS_HIGH.compare(o, value) >= 0 ? true : false);
+          o -> NullSafeComparator.NULLS_HIGH.compare(o, value) >= 0);
     }
 
     public Predicate<Object> matches(Pattern pattern) {
 
-      return new io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.ValueComparingPredicate(
+      return new DaprPredicateQueryCreator.ValueComparingPredicate(
           part.getProperty(), null, o -> {
         if (o == null) {
           return false;
@@ -206,7 +206,7 @@ public class PredicateQueryCreator extends AbstractQueryCreator<KeyValueQuery<Pr
     }
 
     public Predicate<Object> matches(Object value) {
-      return new io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.ValueComparingPredicate(
+      return new DaprPredicateQueryCreator.ValueComparingPredicate(
           part.getProperty(), value, o -> {
 
         if (o == null || value == null) {
@@ -227,7 +227,7 @@ public class PredicateQueryCreator extends AbstractQueryCreator<KeyValueQuery<Pr
     }
 
     public Predicate<Object> in(Object value) {
-      return new io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.ValueComparingPredicate(
+      return new DaprPredicateQueryCreator.ValueComparingPredicate(
           part.getProperty(), value, o -> {
 
         if (value instanceof Collection<?> collection) {
@@ -247,7 +247,7 @@ public class PredicateQueryCreator extends AbstractQueryCreator<KeyValueQuery<Pr
 
     public Predicate<Object> contains(Object value) {
 
-      return new io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.ValueComparingPredicate(
+      return new DaprPredicateQueryCreator.ValueComparingPredicate(
           part.getProperty(), value, o -> {
 
         if (o == null) {
@@ -263,7 +263,7 @@ public class PredicateQueryCreator extends AbstractQueryCreator<KeyValueQuery<Pr
         }
 
         if (o instanceof Map<?, ?> map) {
-          return map.values().contains(value);
+          return map.containsValue(value);
         }
 
         if (value == null) {
@@ -281,7 +281,7 @@ public class PredicateQueryCreator extends AbstractQueryCreator<KeyValueQuery<Pr
     }
 
     public Predicate<Object> startsWith(Object value) {
-      return new io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.ValueComparingPredicate(
+      return new DaprPredicateQueryCreator.ValueComparingPredicate(
           part.getProperty(), value, o -> {
 
         if (!(o instanceof String s)) {
@@ -298,7 +298,7 @@ public class PredicateQueryCreator extends AbstractQueryCreator<KeyValueQuery<Pr
     }
 
     public Predicate<Object> endsWith(Object value) {
-      return new io.diagrid.spring.core.keyvalue.repository.PredicateQueryCreator.ValueComparingPredicate(
+      return new DaprPredicateQueryCreator.ValueComparingPredicate(
           part.getProperty(), value, o -> {
 
         if (!(o instanceof String s)) {
@@ -329,7 +329,8 @@ public class PredicateQueryCreator extends AbstractQueryCreator<KeyValueQuery<Pr
 
     /**
      * Creates a new {@link ValueComparingPredicate}.
-     * @param path The path to the property to compare.
+     *
+     * @param path  The path to the property to compare.
      * @param value The value to compare.
      * @param check The function to check the value.
      */
@@ -355,4 +356,3 @@ public class PredicateQueryCreator extends AbstractQueryCreator<KeyValueQuery<Pr
   }
 
 }
-
